Config packages reorganized (#27)
