+++ cmd/honeypot/config/broadcast.go
@@ -1,0 +1,23 @@
+package config
+
+import "github.com/kyberorg/honeypot/cmd/honeypot/util"
+
+var (
+	loginAttemptBroadcaster *util.LoginAttemptBroadcaster
+)
+
+func init() {
+	once.Do(func() {
+		//init broadcaster
+		initLoginAttemptBroadcaster()
+	})
+}
+
+func GetLoginAttemptBroadcaster() *util.LoginAttemptBroadcaster {
+	return loginAttemptBroadcaster
+}
+
+func initLoginAttemptBroadcaster() {
+	loginAttemptBroadcaster = util.NewLoginAttemptBroadcaster()
+	go loginAttemptBroadcaster.Start()
+}
