+++ cmd/honeypot/config/config.go
@@ -1,7 +1,6 @@
 package config
 
 import (
-	"github.com/kyberorg/honeypot/cmd/honeypot/util"
 	"gopkg.in/alecthomas/kingpin.v2"
 	"sync"
 )
@@ -37,9 +36,8 @@ var (
 
 //internal vars
 var (
-	once                    sync.Once
-	appConfig               *applicationConfiguration
-	loginAttemptBroadcaster *util.LoginAttemptBroadcaster
+	once      sync.Once
+	appConfig *applicationConfiguration
 )
 
 //applicationConfiguration application configuration values
@@ -79,8 +77,6 @@ func init() {
 	once.Do(func() {
 		//parse flags
 		kingpin.Parse()
-		//init broadcaster
-		initLoginAttemptBroadcaster()
 	})
 
 	appConfig = &applicationConfiguration{
@@ -107,16 +103,7 @@ func GetAppConfig() *applicationConfiguration {
 	return appConfig
 }
 
-func GetLoginAttemptBroadcaster() *util.LoginAttemptBroadcaster {
-	return loginAttemptBroadcaster
-}
-
 //IsPromMetricsModuleEnabled says if PromMetrics module is enabled or not, based on activation flag.
 func IsPromMetricsModuleEnabled() bool {
 	return appConfig.PromMetrics.Enabled
-}
-
-func initLoginAttemptBroadcaster() {
-	loginAttemptBroadcaster = util.NewLoginAttemptBroadcaster()
-	go loginAttemptBroadcaster.Start()
 }
